Move YAML rule definitions into index.js
This PR moves our rule definitions from a YAML file into index.js itself.

ESLint rule configuration examples are broadly provided in JSON format,
not YAML. Moreover, while one generally alredy knows JSON if familiar
with YAML, the opposite is not always true. Rather than requiring an
additional parsing library and making the barrier to entry for potential
contributors higher, I'm proposing we use the JS file exports for our
rule definitions instead.

I considered moving to JSON, but the simplicity of having our JS rules
defined in JS (where we can comment, process and format as normal) felt
best. The `.eslintrc` in the project now just points back to index.js.

The rules should not have been altered in their transition from YAML to
JS, _with the exception_ of jsx-curly-spacing where the `children` rule
had to be removed because it is not available in the declared peerDep
version of eslint-plugin-react. I will reinstate it in a dependency
upgrade PR.