Allow use of filesystem functions
For I think not good reasons, the use of fopen, etc are not recommended, and WP_Filesystem is instead recommended. I don't agree with this, and recommend we should instead encourage them. The filesystem functions provide a great abstraction for reading and writing streams, and WP_Filesystem on the other hand... well yeah.