Swap `jsx-a11y/href-no-hash` for `jsx-a11y/anchor-is-valid`
`href-no-hash` has been replaced by a more comprehensive `anchor-is-valid`
rule, but the React App ESLint config has not yet updated to support this
recent version of the a11y ESLint rules. We want the more robust behavior,
so we manually disable the removed rule (with a comment to revisit later)
and manually specify the new rule in our config.